teleop_requirements.txt
absl-py==2.3.1
aioice==0.10.2
annotated-types==0.7.0
asyncio==3.4.3
attrs==25.4.0
av==11.0.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.61.0
google-crc32c==1.7.1
h264-profile-level-id==1.0.0
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
jax==0.8.1
jaxlib==0.8.1
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mediapipe==0.10.14
ml_dtypes==0.5.4
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.9.0.80
opt_einsum==3.4.0
packaging==25.0
pillow==10.4.0
protobuf==4.25.8
pycparser==2.23
pycryptodome==3.20.0
pydantic==1.10.24
pydantic_core==2.41.5
pyee==11.1.1
PyJWT==2.10.1
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
scipy==1.16.3
sdp-transform==1.1.0
six==1.17.0
sounddevice==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
videosdk==0.0.2
vsaiortc==0.0.8
websockets==12.0
Werkzeug==3.1.4
